+++ server/src/modules/recipes/recipesRepository.ts
@@ -1,27 +1,5 @@
 import databaseClient, { type Rows } from "../../../database/client";
-
-interface RecipeBase {
-  id: number;
-  name: string;
-  image: string;
-}
-
-export type AvailableFilters =
-  | "name"
-  | "price"
-  | "duration"
-  | "usersRanking"
-  | "ecoRanking"
-  | "page";
-
-export interface RecipesParams {
-  name?: string;
-  price?: string;
-  duration?: string;
-  usersRanking?: string;
-  ecoRanking?: string;
-  page?: string;
-}
+import type { RecipeBase, RecipesParams } from "../../types/express/recipe";
 
 class RecipesRepository {
   async readSearchRecipes(recipesParams: RecipesParams) {
