+++ server/src/types/express/recipe.ts
@@ -1,0 +1,22 @@
+export interface RecipeBase {
+  id: number;
+  name: string;
+  image: string;
+}
+
+export type AvailableFilters =
+  | "name"
+  | "price"
+  | "duration"
+  | "usersRanking"
+  | "ecoRanking"
+  | "page";
+
+export interface RecipesParams {
+  name?: string;
+  price?: string;
+  duration?: string;
+  usersRanking?: string;
+  ecoRanking?: string;
+  page?: string;
+}
